Add the function initscaling().
The function initscaling() computes the vectors of logarithm sums and
nonzero count for both rows and columns. These are used throughout the
computations to find the row and column factors.

Also rewrite scalefactor() to use the expression used in ititscaling()
to make clear that they compute the logarithm in the same way.