Check that the number of entries of each line of the mps file.
Raise an IndexError exception if the number of entries in a line of the mps
file does not correspond to the number of entries expected for that section.
In this case, interrupt the reading immediately, reporting the culprit line.

This may occur if the entries separators (COLUMNS and RHS) don't appear in
the right position, or just because a line is improperly formatted.