Compute the duality gap inside xi().
This avoids having the computation of the duality gap spread around. In two
cases, this was happening already at the same time when xi() was called, so
this is only a clean up.

In solver() the logic has to be changed from before, since self.gap would
give the new duality gap. So we have to store the previous gap and reverse
the check.