Modify init() to ensure that the initial vectors are not zero.
Inside init(), it is possible that the first estimates of x or s are zero.
In such case, the quantities dp or dd would be zero, which means that the
vectors are not shifted away from the zero axis, and dividing by sum(x) or
sum(s) would generate NaNs.

This change makes sure that the vectors are bounded away from zero.