Stop using matrices to store vectors.
We can gain in speed and code clarity by using arrays instead of matrices for
vectors. Some imports from numpy can be removed, and now we can use diag()
instead of diagflat(), since the vectors are already flat.

With this change, the solution time for adlittle goes from 2.5s to 1.5s,
while the time for 16 iterations of fxm goes from 1m29s to 1m25s.