Be more clever in building the matrix D in the normalequations class.
It's possible to save a few operations by generating the elements of the
diagonal matrix D by doing an element-wise division rather than doing an
expensive matrix by matrix multiplication and computing an inverse.

With this change, the solution time for adlittle goes from 1.5s to 1.2s,
while the time for 16 iterations of fxm goes from 1m25s to 49.1s.