Optimize the use of arrays in the normalequations class.
In the normalequations class, we store only the x and d arrays of diagonal
elements instead of the (dense) diagonal matrices. Thus, we can perfom all
operations without using matrix methods. In particular we can do elementwise
divisions rather than multiplications by the inverse.

As a byproduct, more imports from numpy can be removed, and the code is now
much cleaner. No more asmatrix() and ravel()!

With this change, the solution time for adlittle goes from 1.2s to 0.5s,
while the time for 16 iterations of fxm goes from 49.1s to 1.6s.