Change __parseRhs() to work when no rhs name has been given.
Until now we assumed that the rhs name is given. However, we should cope
also when it is not given. To do this, we take the row-column pairs from
the end of the line, remove them from the line, and stop when less than two
elements are left.

This allows to solve blend.