Catch an exception from stepsize() if there are no negative ratios.
If the search direction has no negative components, then all the ratios are
positive, and the expression
   min(ratios[ratios > 0.0])
will raise a ValueError exception.

This case can arise in unbounded problems. For the time being we return a
stepsize of zero, but in the future we should deal with unboundedness in
a better way.