Ensure that we remove all empty rows from the right-hand side.
If there are empty rows in the last positions, these were not removed from
the right-hand side vector. To find out if this is the case, we work out
the number of rows that should have been removed, and delete them at the
end of the rhs vector.

This problem was seen in brandy.