Rework the normalequations class to prepare for upper bounds.
This introduces the self.D member, which will then be expanded to take into
account for the upper bound contribution, so to stop abusing self.d.

The double division necessary to produce self.D regresses 25fv47, which now
reports a failure, even though it gets very close to the solution.