Adjust rowNames and rowTypes in __deleteEmptyRows().
The easiest way to accomplish this is to ensure that rowNames maintains the
insertion order, which is not guaranteed by a standard dictionary but it is
by an OrderedDict. With this change, in __deleteEmptyRows() we can check if
a row index corresponds to an empty row and remove the corresponding key
from rowNames and rowTypes if that is the case.

To make sure that also the right-hand side vector is updated correctly, we
create it as a sparse dictionary, so that removing the elements corresponding
to empty rows is easy.

This change is necessary to allow moving the addition of slacks from the Mps
to the Hippy class.